+++ examples/src/main/kotlin/latexLabels.kt
@@ -1,29 +1,7 @@
-import kotlinx.html.script
-import kotlinx.html.unsafe
 import kscience.plotly.*
 
-val customMathJaxHeader = HtmlFragment {
-    script {
-        type = "text/x-mathjax-config"
-        unsafe {
-            //language=JavaScript
-            +"""
-            MathJax.Hub.Config({
-                tex2jax: {inlineMath: [['$','$'], ['\\(','\\)']]}
-            });
-            """
-        }
-    }
-    script {
-        type = "text/javascript"
-        async = true
-        src = "https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.7/MathJax.js?config=TeX-MML-AM_SVG"
-    }
-}
-
-
 fun main() {
-    Plotly.page(customMathJaxHeader, cdnPlotlyHeader) {
+    Plotly.page(mathJaxHeader, cdnPlotlyHeader) {
         plot {
             scatter {
                 x(2, 3, 4, 5)