+++ CHANGELOG.md
@@ -7,12 +7,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+## [0.3.1]
+### Added
 - Table widget implementation by @ArtificialPB
+- Mathjax header promoted to stable
+- Tabbed plots layout (experimental)
+- Trace value builders for functions and ranges (experimental)
 
 ### Changed
 - **Breaking API change!** Trace `text` replaced by `TraceValues`
 - Moved to DataForge 0.3 API
-- Kotlin 1.4.21
+- Kotlin 1.4.30
+- **JVM-IR**  
 - Plot `Config` moved to constructor
 - Replaced direct color accessor by a delegate
 
@@ -22,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - https://github.com/mipt-npm/plotly.kt/issues/53
+- Add JQuery to Bootstrap headers
 
 ### Security